ItemController (Adam) (#21)
* ItemController

Created a class that, at least I hope, initializes RSSClient and creates an ObservableList, retrievable with getObservableList().  This is a work in progress but I am very sleepy now and want to commit my progress.

Also updated class Item to use StringProperty instead of simple String attributes.  The functionality of the annotated set methods are unchanged, probably.

* added initialization of properties, fixed a bug in my first commit